Move flatten_byte_* to lowering
The operations in this code are not tied to Boolean flattening and instead do
rewrites that are usable within any back-end.
