+++ src/solvers/lowering/expr_lowering.h
@@ -11,8 +11,18 @@ Author: Michael Tautschnig
 
 #include <util/expr.h>
 
+class byte_extract_exprt;
+class byte_update_exprt;
 class namespacet;
 class popcount_exprt;
+
+exprt flatten_byte_extract(const byte_extract_exprt &src, const namespacet &ns);
+
+exprt flatten_byte_update(const byte_update_exprt &src, const namespacet &ns);
+
+exprt flatten_byte_operators(const exprt &src, const namespacet &ns);
+
+bool has_byte_operator(const exprt &src);
 
 /// Lower a popcount_exprt to arithmetic and logic expressions
 /// \param expr  Input expression to be translated
